moved mbot_loop()
1. no compile error, the robot still run testing on mbot classic
2. next: get rid of all the if statement with robot type in mbot_loop()
