+++ src/mbot.h
@@ -19,6 +19,7 @@
 #include <inttypes.h>
 
 #include "mbot_comms.h"
+#include "mbot_odometry.h"
 
 // TODO: Decide which controller is used, open loop = 1, PID = 0
 #define OPEN_LOOP 1
@@ -26,20 +27,24 @@
 // initializing in mbot.c
 extern mbot_bhy_data_t mbot_imu_data;
 
-// initializing in mbot_classic.c or 
+// initializing in mbot_classic.c 
 extern mbot_params_t params;
+extern uint64_t timestamp_offset;
+extern uint64_t global_utime;
+extern uint64_t global_pico_time;
+extern bool global_comms_status;
+extern int drive_mode;
 
 // Global pointer to the i2c bus
 static i2c_inst_t *i2c;
-
 
 //shared functions for omni and classic
 int mbot_init_pico(void);
 int mbot_init_hardware(void);
 void print_mbot_params(const mbot_params_t* params);
 
-bool mbot_loop(repeating_timer_t *rt);
 void mbot_read_encoders(serial_mbot_encoders_t* encoders);
+void mbot_read_imu(serial_mbot_imu_t *imu);
 void mbot_calculate_motor_vel(serial_mbot_encoders_t encoders, serial_mbot_motor_vel_t *motor_vel);
 
 //helper functions
